got it working kind of
So now it does what we intended, if the date is out of range it defaults to the closest date. However this creates a new problem because if the date is out of range for a year we have yearly data on it doesnt let you look at the yearly summary so that's an L. I think the solution here is to make the yearly review request/handle the info differently than the monthly. much to consider. xoxo gossip girl.